test(map): start spec'ing EuropeanaMap component
