+++ packages/components/map/package.json
@@ -35,7 +35,8 @@
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "catalog:vite7",
-    "vite": "catalog:vite7"
+    "vite": "catalog:vite7",
+    "@vue/test-utils": "catalog:vue3"
   },
   "peerDependencies": {
     "vue": "catalog:vue3"
