+++ packages/components/map/vite.config.js
@@ -11,6 +11,7 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": fileURLToPath(new URL("./src", import.meta.url)),
+      "@test": fileURLToPath(new URL("./test", import.meta.url)),
     },
   },
 });
