+++ pnpm-workspace.yaml
@@ -31,7 +31,8 @@ catalogs:
     "@vitejs/plugin-vue": "6.0.7"
     vite: "7.3.5"
   vue3:
-    vue: 3.5.38
+    vue: "3.5.38"
+    "@vue/test-utils": "2.4.11"
     "@vueuse/core": "14.3.0"
 packages:
   - packages/**
